msg/Pipe: do not deliver delayed message if disconnected
Normally when we mark_down a connection, we set the pipe state to CLOSED
and prevent any further messages from coming through.  If the delayed
delivery is enabled, we may have more messages delayed at the time of the
mark_down.  Later, when we are finished delaying, we will incorrectly
deliver them and hit asserts like

  -118> 2014-08-28 23:40:51.230851 7f9a5fe99700 -1 osd/OSD.cc: In function 'virtual void OSD::ms_fast_dispatch(Message*)' thread 7f9a5fe99700 time 2014-08-28 23:40:51.229339
osd/OSD.cc: 5501: FAILED assert(session)

 ceph version 0.84-823-gf807a7e (f807a7ee2cab3d3eb0877e4846580fdbf3a6b3d0)
 1: (ceph::__ceph_assert_fail(char const*, char const*, int, char const*)+0x8b) [0xb7a60b]
 2: (OSD::ms_fast_dispatch(Message*)+0x3db) [0x6988ab]
 3: (DispatchQueue::fast_dispatch(Message*)+0x56) [0xc191e6]
 4: (Pipe::DelayedDelivery::entry()+0x180) [0xc32640]
 5: (()+0x8182) [0x7f9a851c4182]
 6: (clone()+0x6d) [0x7f9a8373038d]

Fix this by only doing the delayed delivery if the connection is still
connected.

Note that this only causes problems when we are injecting artificial
delays; a real cluster would never use any of this code.

Fixes: #9295
Backport: firefly
Signed-off-by: Sage Weil <sage@redhat.com>